# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NEXTAN Costing & Quote Generator API",
    "description": "Internal costing sheets and client-facing branded quote exports.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/costing-sheets/{sheet_id}/tnc-additions": {
      "get": {
        "tags": [
          "T&C Additions"
        ],
        "summary": "List Sheet Tnc",
        "description": "Purpose: List sheet-specific T&C bullet points.\nInputs: sheet_id (str UUID)\nOutputs: list[TncRead]\nOwner: [Claude]",
        "operationId": "list_sheet_tnc_api_v1_costing_sheets__sheet_id__tnc_additions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sheet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TncRead"
                  },
                  "title": "Response List Sheet Tnc Api V1 Costing Sheets  Sheet Id  Tnc Additions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "T&C Additions"
        ],
        "summary": "Create Sheet Tnc",
        "description": "Purpose: Add a T&C bullet to a costing sheet.\nInputs: sheet_id (str UUID), TncCreate\nOutputs: TncRead\nOwner: [Claude]",
        "operationId": "create_sheet_tnc_api_v1_costing_sheets__sheet_id__tnc_additions_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sheet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Sheet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TncCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TncRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tnc-additions/{tnc_id}": {
      "put": {
        "tags": [
          "T&C Additions"
        ],
        "summary": "Update Tnc",
        "description": "Purpose: Update a sheet-level T&C bullet (bullet_point or display_order).\nInputs: tnc_id (str UUID), TncUpdate\nOutputs: TncRead\nOwner: [Claude]",
        "operationId": "update_tnc_api_v1_tnc_additions__tnc_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tnc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tnc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TncUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TncRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "T&C Additions"
        ],
        "summary": "Delete Tnc",
        "description": "Purpose: Delete a sheet-level T&C bullet.\nInputs: tnc_id (str UUID)\nOutputs: 204 No Content\nOwner: [Claude]",
        "operationId": "delete_tnc_api_v1_tnc_additions__tnc_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tnc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tnc Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/settings/tnc-additions": {
      "get": {
        "tags": [
          "T&C Additions"
        ],
        "summary": "List Global Tnc",
        "description": "Purpose: List global T&C additions (shared across all costing sheets on export).\nInputs: none\nOutputs: list[TncRead]\nOwner: [Claude]",
        "operationId": "list_global_tnc_api_v1_settings_tnc_additions_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TncRead"
                  },
                  "type": "array",
                  "title": "Response List Global Tnc Api V1 Settings Tnc Additions Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "T&C Additions"
        ],
        "summary": "Create Global Tnc",
        "description": "Purpose: Create a new global T&C addition.\nInputs: TncCreate\nOutputs: TncRead\nOwner: [Claude]",
        "operationId": "create_global_tnc_api_v1_settings_tnc_additions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TncCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TncRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/v1/settings/tnc-additions/{tnc_id}": {
      "put": {
        "tags": [
          "T&C Additions"
        ],
        "summary": "Update Global Tnc",
        "description": "Purpose: Update a global T&C addition.\nInputs: tnc_id (str UUID), TncUpdate\nOutputs: TncRead\nOwner: [Claude]",
        "operationId": "update_global_tnc_api_v1_settings_tnc_additions__tnc_id__put",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tnc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tnc Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TncUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TncRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "T&C Additions"
        ],
        "summary": "Delete Global Tnc",
        "description": "Purpose: Delete a global T&C addition.\nInputs: tnc_id (str UUID)\nOutputs: 204 No Content\nOwner: [Claude]",
        "operationId": "delete_global_tnc_api_v1_settings_tnc_additions__tnc_id__delete",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "tnc_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Tnc Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TncCreate": {
        "properties": {
          "bullet_point": {
            "type": "string",
            "title": "Bullet Point"
          },
          "display_order": {
            "type": "integer",
            "title": "Display Order",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "bullet_point"
        ],
        "title": "TncCreate",
        "description": "Purpose: Body for POST /.../tnc-additions. Creates a new T&C bullet.\nInputs: bullet_point, display_order\nOutputs: N/A\nOwner: [Claude]"
      },
      "TncRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "bullet_point": {
            "type": "string",
            "title": "Bullet Point"
          },
          "display_order": {
            "type": "integer",
            "title": "Display Order"
          },
          "costing_sheet_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Costing Sheet Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "bullet_point",
          "display_order",
          "created_at",
          "updated_at"
        ],
        "title": "TncRead",
        "description": "Purpose: T&C Addition record in list/get responses. Used for both sheet-level\n         (CostingSheetTncAddition) and global (GlobalTncAddition).\nInputs: N/A (response body)\nOutputs: id, bullet_point, display_order, costing_sheet_id (None for global), timestamps\nOwner: [Claude]"
      },
      "TncUpdate": {
        "properties": {
          "bullet_point": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bullet Point"
          },
          "display_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Order"
          }
        },
        "type": "object",
        "title": "TncUpdate",
        "description": "Purpose: Body for PUT .../tnc-additions/{id}. Updates a T&C bullet.\nInputs: bullet_point, display_order (both optional)\nOutputs: N/A\nOwner: [Claude]"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}